requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
pandas==1.3.3
werkzeug==2.0.1
openpyxl==3.0.9
xlrd==2.0.1  # For old Excel format support